Bespoke Sockets (#19)
This change splits the sockets into three distinct kinds.

This ended up being a bit messier (with generics) than I had hoped. The goal of this is to allow different behaviors for different kinds of sockets, as well as spending less time iterating over "wrong" sockets when processing incoming messages, which is a fairly hot part of the system.