Back to a single ergot crate (#86)
This PR essentially undoes https://github.com/jamesmunns/ergot/pull/22, as I don't plan to necessarily support a non-integrated net stack world.

It's very possible some day I will decide to revert this revert, but for now, this seems to be a pragmatic way to reduce some of the friction that I've been struggling with.