Fix compiler warnings (#169)
* Remove unused imports

* Silence warning about unused `info!` macro

This macro is only used by code under the `tokio-std` feature, but is
available under more general features. An alternative way to address
this warning, would be to only make the macro available under
`tokio-std`, but then the re-export below would fail.

Since that re-export is annotated with `#[allow(unused_imports)]`, I
conclude that those macros are meant to always be available, even if
they are only used intermittently. I think this makes sense for logging
macros.

Given all that, the `#![allow(unused_macros)]` which I'm adding here,
seems like the right measure.