Fix warnings in demo code; configure CI to fail, should any reappear (#170)
* Remove unused variables in demos

* Remove unused imports in demos

* Remove unused functions in demo code

Also removed the commented code that mentioned those unused functions.

* Remove more unused code in tests

* Replace per-step `RUSTFLAGS` with global one

This should not make any difference for this workflow, since all steps
already set the flag. But it should make the workflow easier to maintain
going forward. I think failing the CI build on warnings is always a good
idea, and now we don't have to remember that for each new step.

* Configure Cargo to output color in Clippy CI build

This makes the output much friendlier, if there are warnings or errors.

* Replicate Clippy build env in all workflows

This makes Cargo output nicer and fails all builds in case of warnings.

There might be better ways to do this instead of duplicating this
configuration everywhere, but I don't know. Maybe there are ways to
share configuration between workflows, or maybe it makes sense to merge
all the jobs into one workflow, where they may share this environment.

GitHub Actions is definitely not my specialty, and I don't want to
accidentally mess up any of this setup's good qualities. So I opted to
just duplicate this configuration in all workflows.